requirements.txt
beautifulsoup4==4.11.1
psycopg2-binary==2.9.3
pyTelegramBotAPI==4.6.0
requests==2.23.0
